requirements/project.txt
Mezzanine==1.4.16
PIL==1.1.7
South==0.8.2
TwitterSearch==0.78.3
argparse==1.2.1
django-autoslug==1.7.1
django-social-auth==0.7.28
djangorestframework==2.3.9
httplib2==0.8
oauth2==1.5.211
oauthlib==0.6.0
psycopg2==2.5.1
python-instagram==0.8.0
requests==2.0.1
requests-oauthlib==0.4.0
simplejson==3.3.1
tweepy==2.1
twitter==1.10.2
wsgiref==0.1.2